Simplify integration with reporting tools
We use Raygun to log our javascript errors, and we'd like to automatically report migration warnings there too. Raising an event when a warning is logged seems a very natural way to do this - we can come scrape $.migrateWarnings when triggered. A stack trace also seems like a very useful thing to have, so extending the warning seems like a natural way to achieve this (it also helps a lot with Raygun, which likes to send Error objects).