Traces: commas replaced by semi-columns
Standardization of the trace text shape to facilitate textual search into https://github.com/jquery/jquery-migrate/blob/1.x-stable/warnings.md
I arbitrary chose the semi-column as a sentence separator. This is not important, the goal is to be homogeneous between the sources and the documentation, because there is at least one inconsistency with "JQMIGRATE: jQuery.parseJSON is deprecated; use JSON.parse", that I want to fix.