Data: Support object as second argument
jQuery.data supports passing an object as a second argument and sets data
for each entry in that object. This warns the user about each violating
non-camelCase key in the object. This also adds a type check to avoid
a current bug where jQuery.camelCase throws an error because a non-string
is passed to it.

Fixes #198
Closes #235