Tests: Remove unneeded static HTML markup in fixture
This was leftover from 1.x and never culled. Newer tests should in most cases
just insert the markup they need to do their tests and not assume static markup
in the main fixture. For more complex cases it may be better to create an iframe
test rather than cluttering the common fixture in index.html!