Core: Support jQuery slim build
While it'd be hard for Migrate to support all the possible jQuery builds,
the slim build is special as it's officially maintained by the jQuery team,
published to npm & uploaded to CDNs.

This commit adds support for the jQuery slim build in jQuery 3.0.0 & newer.
The slim build is now tested next to the regular one.

Fixes #345