CSS: Don't warn for number values when prop in jQuery.cssNumber
Relying on numerical values for most CSS properties having `"px"`
auto-appended when set via `.css( prop, value )` is deprecated. So far,
Migrate was warning for number values for all CSS properties with the
exception of the ones to which jQuery 4.x will still auto-append `"px"`.
However, if the used jQuery version contains the property as a key in
`jQuery.cssNumber`, it will already skip the `"px"` auto-appending behavior,
exactly as jQuery 4.x will. This commit removes a warning in such cases.

This change will help jQuery UI from not triggering Migrate warnings.

Closes gh-348