CSS: Don't warn against `.css( "z-index", numberValue )` in jQuery >=4
When used against jQuery Git (future jQuery 4), there were warnings when
`.css()` was used against `z-index` with a number value. This didn't make sense
as `z-index` never had `px` auto-appended. Commit
d25ecd68777e4be1185b61e07eb81dce2cfd56d6 was supposed to solve a similar problem
but it only worked fine for `3.x`.

To make the warning not show up when used with jQuery >=4 we need to fill in
`jQuery.cssNumber` with the latest version from jQuery 3.x so that the same
properties are allowed to have number values used with `.css()`.

Fixes gh-438