Tests: Add a new esmodules flag, remove the raw one
Migrating from AMD to ES modules in gh-343 made the `raw` testing mode break as
it relied on loading raw source files. This change removes the `raw` mode in
favor of the new `esmodules` one.

Setup is different than in jQuery as jQuery uses AMD to load tests and Migrate
just loads test files directly.

When Migrate is loaded via its source ES modules, test files are loaded as
modules as well. This is necessary as module scripts behave as if loaded with
the `defer` attribute so tests need to be deferred as well.

Iframe tests are skipped in `esmodules` mode as their setup is incompatible
with the above considerations.

In addition to the above, Karma tests now also run in esmodules mode for easier
verification this mode is not broken by PRs.

Closes gh-468
Ref gh-343