Build: Make Migrate properly recognized as a CommonJS module in Node.js
PR gh-512 added the `"type": "module"` field to the top-level `package.json`,
making Node-based workflows fail when trying to require Migrate as it's not
exposed as ESM. To fix this, add a small `package.json` with just
`"type": "commonjs"`. Also, add Node.js smoke tests as the simplest way to
verify this change.

Fixes gh-523
Ref gh-512