Docs: Improve the solution for `boolean-attributes`
Because jQuery 4.0.0 will not transform attribute values in the getter, the
following code changes between jQuery 3.x & 4.x:

```js
$( elem ).prop( "disabled", true );
$( elem ).attr( "disabled" );
```

In jQuery 3.x, the reported attribute value is `"disabled"`, in jQuery 4.x it
will be an empty string. This may create issues when migrating so Migrate
patches and warns against it.

Add a recommendation to avoid using the `.attr()` getter on boolean attrs.

Closes gh-588